requirements.txt
flask
flask_socketio
torch
transformers
llama-cpp-python
piper-tts

sounddevice
pydub
espeak_phonemizer

wave
numpy
ollama

torchaudio
langchain

openai==0.28.1